requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
tqdm>=4.62.0
torch>=2.0.0
torchvision>=0.15.0
scipy>=1.7.0
Pillow>=8.0.0
folium>=0.12.0 